+++ libs/logger/src/index.ts
@@ -1,4 +1,3 @@
 export * from './lib/config/logger';
-export * from './lib/interceptors/logging.interceptor';
 export * from './lib/module';
 export * from './lib/utils/with-logger-context.decorator';