+++ libs/logger/src/lib/__tests__/app/config/logger-root.config.ts
@@ -1,0 +1,9 @@
+import { ValidateNested } from 'class-validator';
+import { Type } from 'class-transformer';
+import { LoggerConfig } from '../../../config/logger';
+
+export class LoggerRootConfig {
+  @ValidateNested()
+  @Type(() => LoggerConfig)
+  logger!: LoggerConfig;
+}