+++ libs/file-storage/CHANGELOG.md
@@ -1,4 +1,62 @@
 Softkit Core Libraries Changelog
+## [0.4.3](https://github.com/softkitit/softkit-core/compare/file-storage-0.4.2...file-storage-0.4.3) (2024-07-05)
+
+
+### Bug Fixes
+
+* **file-storage:** fixed cache control header for presign url and changed status to 302 temporary redirect ([31314e1](https://github.com/softkitit/softkit-core/commit/31314e10b5194ec0a1b524f6d28f3b7028b57caa))
+
+## [0.4.2](https://github.com/softkitit/softkit-core/compare/file-storage-0.4.1...file-storage-0.4.2) (2024-07-05)
+
+
+### Bug Fixes
+
+* **file-storage:** fixed cache control header for presign url ([ea177d1](https://github.com/softkitit/softkit-core/commit/ea177d1a65c6df353002247b94aa697ccbd5a207))
+
+## [0.4.1](https://github.com/softkitit/softkit-core/compare/file-storage-0.4.0...file-storage-0.4.1) (2024-07-04)
+
+
+### Bug Fixes
+
+* **file-storage:** download files from folder ([f1e4994](https://github.com/softkitit/softkit-core/commit/f1e499437b422677f888b36c9f0ced7cf0bb7b44))
+
+## [0.4.0](https://github.com/softkitit/softkit-core/compare/file-storage-0.3.0...file-storage-0.4.0) (2024-07-02)
+
+
+### Features
+
+* **file-storage:** renamed the presign.dto.ts file ([8483a2c](https://github.com/softkitit/softkit-core/commit/8483a2c5fafe15610cfd425405f3e1619ff60572))
+* **file-storage:** update the request body in swagger ([194af06](https://github.com/softkitit/softkit-core/commit/194af068faf019cbda48253ccf8383cb3807ad34))
+* **file-storage:** update the request body in swagger ([a6b2ef8](https://github.com/softkitit/softkit-core/commit/a6b2ef8700d2c65882784cab8f298eea6447226f))
+* **file-storage:** updated the swagger api property ([9b5b5a4](https://github.com/softkitit/softkit-core/commit/9b5b5a4bd65f6fbcc468f35c9fd2034c9ed0ea70))
+* **file-storage:** updated the swagger api property ([61c8ad3](https://github.com/softkitit/softkit-core/commit/61c8ad30f6d982df9f310536d5f0386ba11a5475))
+* **file-storage:** updated the swagger api property ([9e3bf50](https://github.com/softkitit/softkit-core/commit/9e3bf505adad3dfdf3df3fc86c9f1c30ea676f4a))
+* **file-storage:** updated the swagger api property, updated the request dto for the upload file endpoint ([d736dfb](https://github.com/softkitit/softkit-core/commit/d736dfb8f8f2cef4f9d42391776093ac1ccb2962))
+* **file-storage:** updated the swagger api property, updated the request dto for the upload file endpoint ([1aea195](https://github.com/softkitit/softkit-core/commit/1aea1957c5dd87ab9e78c9dd124e072e72679fd8))
+
+## [0.3.0](https://github.com/softkitit/softkit-core/compare/file-storage-0.2.0...file-storage-0.3.0) (2024-06-25)
+
+
+### Features
+
+* **file-storage:** moved dependencies from the peer dependencies to the dependencies ([53d6a7a](https://github.com/softkitit/softkit-core/commit/53d6a7a5005f1cbdf5fc0993da5601b4bfd96b9a))
+
+## [0.2.0](https://github.com/softkitit/softkit-core/compare/file-storage-0.1.8...file-storage-0.2.0) (2024-06-24)
+
+
+### Features
+
+* **file-storage:** added the abstract file controller ([3c77cd0](https://github.com/softkitit/softkit-core/commit/3c77cd02a3736e165dd2bd58da6b047e265ef9e7))
+* **file-storage:** added the abstract file controller ([84a10be](https://github.com/softkitit/softkit-core/commit/84a10bec5e45e1d6c4c767f6febcc39be6e64598))
+* **file-storage:** updated after reviewing ([a67fc74](https://github.com/softkitit/softkit-core/commit/a67fc742dc5a849a46be34d0d0e4415bcd7f9a30))
+* **file-storage:** updated after reviewing ([ad4ddfc](https://github.com/softkitit/softkit-core/commit/ad4ddfc85961e2532594e1b4c42afef39cc65d50))
+* **file-storage:** updated after reviewing ([50e0c14](https://github.com/softkitit/softkit-core/commit/50e0c14f1f95ebfa796f8e4495b81011d1da8a4d))
+
+
+### Bug Fixes
+
+* **file-storage:** various small fixed and naming change ([7b48b4e](https://github.com/softkitit/softkit-core/commit/7b48b4e577e9d247e54c5aa637a0f8e027f7f88b))
+
 ## [0.1.8](https://github.com/softkitit/softkit-core/compare/file-storage-0.1.7...file-storage-0.1.8) (2024-03-06)