+++ libs/jobs/CHANGELOG.md
@@ -1,4 +1,26 @@
 Softkit Core Libraries Changelog
+## [0.2.3](https://github.com/softkitit/softkit-core/compare/jobs-0.2.2...jobs-0.2.3) (2024-05-23)
+
+
+### Bug Fixes
+
+* **jobs:** added bullmq as peer deps ([f17a4e7](https://github.com/softkitit/softkit-core/commit/f17a4e765a92fad59bdcc2d39e8c0280ebcde83c))
+
+## [0.2.2](https://github.com/softkitit/softkit-core/compare/jobs-0.2.1...jobs-0.2.2) (2024-05-17)
+
+
+### Bug Fixes
+
+* **jobs:** improved logs to include context and show errors stacktrace properly ([143d769](https://github.com/softkitit/softkit-core/commit/143d7698ad2b6dab83d4c52bf62b88b16f25648b))
+* **jobs:** removed duplicate eslint rule ([bc4f6c1](https://github.com/softkitit/softkit-core/commit/bc4f6c1b893f33f1c25cab86b6d65a042958d123))
+
+## [0.2.1](https://github.com/softkitit/softkit-core/compare/jobs-0.2.0...jobs-0.2.1) (2024-05-11)
+
+
+### Bug Fixes
+
+* **jobs:** updated tests to add logger config after fixing issue with logger misconfiguration ([4e9d316](https://github.com/softkitit/softkit-core/commit/4e9d3163145b0a427b13d335342ada3f7e0be1d1))
+
 ## [0.2.0](https://github.com/softkitit/softkit-core/compare/jobs-0.1.5...jobs-0.2.0) (2024-02-07)